% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Calculate Birthday Probability}
\usage{
birthday(n)
}
\arguments{
\item{n}{The class size of the fucntion}
}
\value{
vector of probabilities
}
\description{
Calculate Birthday Probability
}
\examples{
#Probability for a class of 23 people
birthday(23)

}
