% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data set}
\format{
## `who`
A data frame with 15 rows and 2 columns:
\describe{
\item{DISTANCE}{Distance from fire damage in miles}
\item{DAMAGE}{Fire damage in thousands of dollars}
...
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Science/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
fire
}
\description{
A subset of data from a study conducted in a suburb of a large city of fire damage
}
\keyword{datasets}
