% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{This will create a boxplot where you can change the variables by calling the function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{This is the iterations}

\item{n}{THis is the x-axis}

\item{p}{This is the y-axis}
}
\value{
A boxplot
}
\description{
This will create a boxplot where you can change the variables by calling the function
}
\examples{
\dontrun{mybin(iter100,n =10, p =0.5)}
}
