% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myboot2}
\alias{myboot2}
\title{My Bootstrap Interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{a number of iterations to run}

\item{x}{a numeric vector of data values}

\item{fun}{the name of a function to evaluate x}

\item{alpha}{a significance level}

\item{cx}{a numberic character expansion factor}

\item{...}{further arguments to be passed to or from methods}
}
\value{
a histogram and list of a bootstrap interval of x
}
\description{
My Bootstrap Interval
}
\examples{
\dontrun{myboot2(iter=10000,rnorm(20,10,4),fun="mean",alpha=0.05,cx=1.5)}
}
