% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{My Central Limit Theorem Binomial}
\usage{
mycltb(n, iter, p = 0.5)
}
\arguments{
\item{n}{number of observations}

\item{iter}{number of trials}

\item{p}{probability of success on each trial}
}
\value{
a distribution of the sample mean
}
\description{
My Central Limit Theorem Binomial
}
\examples{
mycltb(n = 20,iter=10000, p = 0.5)
}
