% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My normal curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{a vector of means}

\item{sigma}{a vector of standard deviations}

\item{a}{discrete random variable}
}
\value{
a normal curve of probability
}
\description{
My normal curve
}
\examples{
myncurve(0,1, 0)
}
