% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N = 200, p = 0.95, gamma = 0.02)
}
\arguments{
\item{N}{The number of seats in the flight}

\item{p}{The probability fo the passenger showing up to the flight}

\item{gamma}{probability a plan will be truly overbooked}
}
\value{
a list contatin nd, nc, N, p, and gamma
}
\description{
ntickets
}
