% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Calculate Brithday Probability}
\usage{
birthday(n)
}
\arguments{
\item{n}{Integer or vecotr of itnegers}
}
\value{
vector of probabilities
}
\description{
Calculate Brithday Probability
}
\examples{
#Probbability for a class of 23 people
birthday(23)

}
